pyproject.toml
[project]
name = "agentic-ai-article-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "bedrock-agentcore>=1.1.1",
    "strands-agents>=1.19.0",
    "strands-agents-tools>=0.2.18",
]

[dependency-groups]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.2.5",
]
